+++ src/main/java/org/torproject/onionoo/writer/WeightsDocumentWriter.java
@@ -55,14 +55,14 @@ public class WeightsDocumentWriter implements DocumentWriter {
   private String[] graphNames = new String[] {
       "1_week",
       "1_month",
-      "3_months",
+      "6_months",
       "1_year",
       "5_years" };
 
   private Period[] graphIntervals = new Period[] {
       Period.ofWeeks(1),
       Period.ofMonths(1),
-      Period.ofMonths(3),
+      Period.ofMonths(6),
       Period.ofYears(1),
       Period.ofYears(5) };